// my-weather/src/app/main/model/model.api.ts
// tslint:disable-next-line: no-namespace
export namespace ModelAPI {

	export interface Main {
		temp: number;
		feels_like: number;
		temp_min: number;
		temp_max: number;
		pressure: number;
		humidity: number;
	}

	export interface Wind {
		speed: number;
		deg: number;
	}

	export interface Clouds {
		all: number;
	}

	export interface Sys {
		type: number;
		id: number;
		country: string;
		sunrise: number;
		sunset: number;
	}

	export interface WeatherAPI {
		base: string;
		main: Main;
		visibility: number;
		wind: Wind;
		clouds: Clouds;
		dt: number;
		sys: Sys;
		timezone: number;
		id: number;
		name: string;
		cod: number;
	}

}